chore: fixing config
